feat: Creada tabla preguntas, respuestaAlumno y respuestaPosible. Hemos implementado esas clases para mejorar el hacer preguntas y el tipo test.

+++ backend-api/src/main/java/org/example/backendapi/service/UsuarioService.java
@@ -9,6 +9,7 @@ import org.example.backendapi.mapper.UsuarioMapper;
 import org.example.backendapi.model.dao.IUsuarioDAO;
 import org.example.backendapi.model.entities.TipoRol;
 import org.example.backendapi.model.entities.Usuario;
+import org.example.backendapi.exception.ForbiddenException;
 import org.springframework.stereotype.Service;
 import org.example.backendapi.exception.ResourceNotFoundException;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,6 +23,7 @@ public class UsuarioService {
 
     private final IUsuarioDAO usuarioDAO;
     private final SecurityService securityService;
+    private final JwtService jwtService;
     private final UsuarioMapper usuarioMapper;
 
     @Transactional
@@ -32,6 +34,9 @@ public class UsuarioService {
 
         Usuario profesor = new Usuario();
         profesor.setNombreUsuario(request.getNombreUsuario());
+        profesor.setNombreReal(request.getNombreReal());
+        profesor.setApellidos(request.getApellidos());
+        profesor.setEmail(request.getEmail());
 
         profesor.setHashContrasena(securityService.hashPassword(request.getPasswordPlana()));
         profesor.setFechaCreacion(Instant.now());
@@ -55,7 +60,11 @@ public class UsuarioService {
             throw new BadRequestException("Contraseña incorrecta");
         }
 
-        return usuarioMapper.toResponseDTO(usuario);
+        String token = jwtService.generateToken(usuario);
+        UsuarioResponseDTO responseDTO = usuarioMapper.toResponseDTO(usuario);
+        responseDTO.setToken(token);
+
+        return responseDTO;
     }
 
     public UsuarioResponseDTO buscarUsuarioPorId(Long id) {
@@ -71,10 +80,16 @@ public class UsuarioService {
     }
 
     @Transactional
-    public void borrarUsuario(Long id) {
+    public void borrarUsuario(Long id, Usuario usuarioLogueado) {
         if (!usuarioDAO.existsById(id)) {
             throw new ResourceNotFoundException("No se puede borrar. El usuario con ID " + id + " no existe.");
         }
+
+        if (usuarioLogueado.getRol() == TipoRol.ROL_ESTUDIANTE && !usuarioLogueado.getId().equals(id)) {
+            throw new ForbiddenException("No puedes borrar la cuenta de otro usuario.");
+        }
+        
+        // TODO: Si es profesor, verificar si el alumno le pertenece (opcional)
 
         usuarioDAO.deleteById(id);
     }
